api/requirements.txt
Django>=3.2,<4.0
djangorestframework
pymysql==1.0.2
mysqlclient>=2.0
mysql-connector-python
requests
bs4